add exercise 4.21

+++ part4/blogList/controllers/blogs.js
@@ -18,16 +18,16 @@ blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
-// get all posts
+// post a blog
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
   const { title, url, author, likes } = body;
   if (!title || !url) {
     return response.status(400).end();
   }
-   const token = getTokenFrom(request);
+  const token = getTokenFrom(request);
   // const decodedToken = jwt.verify(request.token, process.env.SECRET);
-   const decodedToken = jwt.verify(token, process.env.SECRET);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
   if (!token || !decodedToken.id) {
     return response.status(401).json({ error: "token missing or invalid" });
   }
@@ -50,8 +50,24 @@ blogsRouter.post("/", async (request, response) => {
 });
 //delete a post
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const token = getTokenFrom(request);
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+  if(!user) {
+    return response.status(401).json({ error: "unauthorized user" });
+  }
+  if (!blog) {
+    return response.status(404).json({ error: "blog not found" });
+  }
+  if (user.id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }
 });
 // update a post
 blogsRouter.put("/:id", async (request, response) => {
